toolkit/components/captchadetection/tests/mochitest/test_cf_turnstile_success.html

<!doctype html>
<html>
  <head>
    <meta charset="utf-8" />
    <title>Cloudflare Turnstile Completion Detection Test</title>
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" href="/tests/SimpleTest/test.css" />
    <script src="GleanTest.js"></script>
    <script src="CaptchaTestingUtils.js"></script>
  </head>
  <body>
    <p id="display"></p>
    <div id="content" style="display: none"></div>
    <pre id="test"></pre>

    <script>
      add_task(async function () {
        await GleanTest.testResetFOG();

        const iframe = await CaptchaTestingUtils.createIframeAndWaitForMessage(
          "cf_turnstile.html",
          document.body,
          "ready"
        );

        const metricSet = CaptchaTestingUtils.waitForMetricSet();
        iframe.contentWindow.postMessage("success", "*");
        await metricSet;

        const puzzleCompletedCount =
          await GleanTest.captchaDetection.cloudflareTurnstileCc.testGetValue();
        is(puzzleCompletedCount, 1, "The puzzle completion count should be 1");

        await CaptchaTestingUtils.clearPrefs();
      });
    </script>
  </body>
</html>
